refactor(embedding): 收窄 token usage 变更到结果返回 API
# 影响范围
- Java/Python `BaseEmbeddingModelConnection` 与 `BaseEmbeddingModelSetup` 保留 `embedWithUsage`/`embed_with_usage` 的结果携带链路；普通 `embed` API 的返回类型不变。
- `BaseVectorStore.query`、自动补 embedding 以及 Python 对应实现恢复仅调用 `embed`，不再进入指标组或计数器。
- OpenAI、Tongyi、Bedrock 的 provider usage 提取及批量聚合能力保持不变；测试改为断言 usage 结果本身。

# 改动影响面
- 主链路：embedding provider -> `EmbeddingResult` -> 调用方显式消费 usage；本提交不再把 usage 写入 `MetricGroup`。
- 向量库扩展点：恢复 `BaseEmbeddingModelSetup.embed` 覆盖语义，并且仅在确有缺失 embedding 时延迟解析模型，避免 Mem0 无模型路径报错。
- 不受影响范围：#860/#861 的 resource metric-group 传播和 action-scoped metric 语义未改动；未引入新的 metric abstraction。

# 功能改进/开发/新增
- 类型: refactor
- 触发条件: 普通资源 API 可并发调用，无法安全保证 `MetricGroup`/`SimpleCounter` 的线程模型；Python vector-store 还绕过了 `embed` 扩展点。
- 根因类别: 指标写入边界不明确、公共 API 扩展点回归、惰性资源解析边界遗漏。
- 行为变化: 有；PR 仅暴露 provider usage 返回 API，不再自动记录 embedding token metrics。

# 验证
- `mvn --batch-mode --no-transfer-progress -pl api -Dtest=BaseEmbeddingModelSetupEmbeddingResultTest test` PASS
- `mvn --batch-mode --no-transfer-progress -pl integrations/embedding-models/bedrock -am -Dtest=BedrockEmbeddingModelTest -Dsurefire.failIfNoSpecifiedTests=false test` PASS
- `uv run --python 3.12 --extra test pytest ...vector_stores... -q` PASS (49 passed, 4 skipped)
- `./tools/lint.sh --check` PASS
- `uv run --python 3.12 --extra lint ruff format --check ... && uv run --python 3.12 --extra lint ruff check ...` PASS
- `git diff --check` PASS
